Gosund EP2: Switch LEDs
With the current template the blue LED was used to indicate on/off state. I switched the two LEDs so that GPIO0 is LedLink_i now and GPIO2 is Led_i and now the green LED is used.

The template looks quite different, so not sure if I did something wrong or if the is some kind of new format.